Moved reply code to its own service, it now creates a reply record and puts the photo filepath in it
